Move identifiers into own mod.
