+++ libs/prisma-models/src/lib.rs
@@ -12,6 +12,7 @@ mod internal_data_model;
 mod model;
 mod order_by;
 mod prisma_value_ext;
+mod projections;
 mod record;
 mod relation;
 mod selected_fields;
@@ -32,6 +33,7 @@ pub use internal_data_model::*;
 pub use model::*;
 pub use order_by::*;
 pub use prisma_value_ext::*;
+pub use projections::*;
 pub use record::*;
 pub use relation::*;
 pub use selected_fields::*;
