+++ libs/prisma-models/src/prelude.rs
@@ -6,6 +6,7 @@ pub use crate::index::*;
 pub use crate::internal_data_model::*;
 pub use crate::model::*;
 pub use crate::order_by::*;
+pub use crate::projections::*;
 pub use crate::record::*;
 pub use crate::relation::*;
 pub use crate::selected_fields::*;
