+++ libs/prisma-models/src/projections/mod.rs
@@ -1,0 +1,5 @@
+mod model_projection;
+mod record_projection;
+
+pub use model_projection::*;
+pub use record_projection::*;
